backend/case-recommendation-engine/requirements.txt
# ── HuggingFace ──────────────────────────────────────────────
datasets>=2.19.0               # Load HF dataset (NEW)
huggingface-hub>=0.23.0

# ── Core Python ──────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0

# ── Machine Learning (scikit-learn) ──────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0

# ── Deep Learning (TensorFlow / Keras) ───────────────────────
tensorflow>=2.13.0

# ── NLP & Transformers ───────────────────────────────────────
nltk>=3.8.1
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.0.0

# ── File Parsing ─────────────────────────────────────────────
PyMuPDF>=1.24.0
python-docx>=1.1.0

# ── Web API ──────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Visualization ────────────────────────────────────────────
matplotlib>=3.7.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0
